Add files via upload
This is the library the dynamic load model used in the provided three full-replica dynamic models for the three versions of the IEEE 39-bus power systems. Please add this library model in your MATLAB library in order to successfully load the dynamic load model when you run the models of the IEEE 39-bus power systems.